+++ Sources/Sentry/include/SentryPrivate.h
@@ -11,6 +11,7 @@
 // otherwise it wont compile
 #import "SentryDateUtil.h"
 #import "SentryDisplayLinkWrapper.h"
+#import "SentryFileIOTracker.h"
 #import "SentryLevelHelper.h"
 #import "SentryLogC.h"
 #import "SentryMeta.h"